CBP - stock Transactions - cash Money Restock - QA Fix GetManager #3012
